fix `Cannot read properties of null (reading 'error')`

+++ app/models/comment.js
@@ -179,20 +179,17 @@ exports.delete = function (p, cb) {
         cb && cb({ error: comment ? comment.error : 'comment not found' });
         return;
       }
-      postModel.fetchPostById(
-        comment.pId,
-        (post = { error: 'post not found' }) => {
-          if (post.error) {
-            cb && cb(post);
-            return;
-          }
-          if (p.uId != post.uId && comment.uId != p.uId) {
-            cb && cb({ error: 'you are not allowed to delete this comment' });
-            return;
-          }
-          getCol().deleteOne(q, combineResult(cb));
-        },
-      );
+      postModel.fetchPostById(comment.pId, (post) => {
+        if (!post || post.error) {
+          cb && cb({ error: post ? post.error : 'post not found' });
+          return;
+        }
+        if (p.uId != post.uId && comment.uId != p.uId) {
+          cb && cb({ error: 'you are not allowed to delete this comment' });
+          return;
+        }
+        getCol().deleteOne(q, combineResult(cb));
+      });
     }),
   );
 };
